pyproject.toml
[tool.poetry]
name = "tlmr-digest"
version = "0.0.1"
description = "More information coming soon"
authors = ["Daniel Burger <public@danielburger.online>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tlmr_digest"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.1"
wget = "^3.2"
pdfplumber = "^0.7.6"
numpy = "^1.24.1"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.12.0"
pylint = "^2.15.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
